move atoms to components folder
