+++ apps/nameguard.io/src/app/components/atoms/NGSearchIcon.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { MagnifyingGlassIcon } from "@/app/atoms/icons/MagnifyingGlassIcon";
+import { MagnifyingGlassIcon } from "@/app/components/atoms/icons/MagnifyingGlassIcon";
 import { useSearchStore } from "@namehash/nameguard-react";
 
 export function NGSearchIcon() {
