+++ apps/nameguard.io/src/app/components/atoms/icons/CheckShieldGrayOutline.tsx
