+++ apps/nameguard.io/src/app/components/atoms/icons/CloudOutlineIcon.tsx
