+++ apps/nameguard.io/src/app/components/atoms/icons/CoinbaseIcon.tsx
