+++ apps/nameguard.io/src/app/components/atoms/icons/CopyIcon.tsx
