+++ apps/nameguard.io/src/app/components/atoms/icons/EmailIcon.tsx
