+++ apps/nameguard.io/src/app/components/atoms/icons/FarcasterLogo.tsx
