+++ apps/nameguard.io/src/app/components/atoms/icons/FigmaIcon.tsx
