+++ apps/nameguard.io/src/app/components/atoms/icons/FileIcon.tsx
