+++ apps/nameguard.io/src/app/components/atoms/icons/FontIcon.tsx
