+++ apps/nameguard.io/src/app/components/atoms/icons/GearWheelIcon.tsx
