+++ apps/nameguard.io/src/app/components/atoms/icons/GithubIconSmall.tsx
