+++ apps/nameguard.io/src/app/components/atoms/icons/IntegrationIcon.tsx
