+++ apps/nameguard.io/src/app/components/atoms/icons/LensProtocolIcon.tsx
