+++ apps/nameguard.io/src/app/components/atoms/icons/MagnifyingGlassIcon.tsx
