+++ apps/nameguard.io/src/app/components/atoms/icons/NameGuardIconLarge.tsx
