+++ apps/nameguard.io/src/app/components/atoms/icons/NameGuardIconSmall.tsx
