+++ apps/nameguard.io/src/app/components/atoms/icons/OpenSeaIcon.tsx
