+++ apps/nameguard.io/src/app/components/atoms/icons/RaribleIcon.tsx
