+++ apps/nameguard.io/src/app/components/atoms/icons/RedirectIcon.tsx
