+++ apps/nameguard.io/src/app/components/atoms/icons/RoadmapNegativeShield.tsx
