+++ apps/nameguard.io/src/app/components/atoms/icons/ShoppingCartIcon.tsx
