+++ apps/nameguard.io/src/app/components/atoms/icons/WalletIcon.tsx
